app/requirements.txt
streamlit>=1.35
pandas>=2.0
plotly>=5.18
folium>=0.15
streamlit-folium>=0.18
jsonschema>=4.17
google-generativeai>=0.8
